requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.2.1
email-validator==2.1.1
pytest==8.1.1
httpx==0.27.0
slowapi==0.1.9
python-multipart==0.0.9
celery==5.3.6
redis==5.0.4
sentry-sdk==1.45.0
python-json-logger==2.0.7
